backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
PyJWT[crypto]>=2.8.0,<3.0
snowflake-snowpark-python>=1.11.0,<2.0
python-dotenv>=1.0.0
